+++ scripts/extensions/markForUser/src/extension.tsx
@@ -1,103 +1,18 @@
-import {ISuperdesk, IExtension, IArticleAction, IArticle, IExtensionActivationResult} from 'superdesk-api';
-import {getMarkForUserModal} from './get-mark-for-user-modal';
-import {uniq} from 'lodash';
+import {ISuperdesk, IExtension, IExtensionActivationResult} from 'superdesk-api';
 import {getDisplayMarkedUserComponent} from './show-marked-user';
-import {manageMarkedUserForSingleArticle} from './managed-marked-user';
+import {getActionsInitialize} from './get-article-actions';
+import {getActionsBulkInitialize} from './get-article-actions-bulk';
 
 const extension: IExtension = {
     activate: (superdesk: ISuperdesk) => {
-        const {gettext} = superdesk.localization;
-        const {isPersonal} = superdesk.entities.article;
-
         const result: IExtensionActivationResult = {
             contributions: {
                 articleListItemWidgets: [getDisplayMarkedUserComponent(superdesk)],
                 authoringTopbarWidgets: [getDisplayMarkedUserComponent(superdesk)],
                 entities: {
                     article: {
-                        getActions: (articleNext) => {
-                            if (isPersonal(articleNext)) {
-                                return Promise.resolve([]);
-                            }
-
-                            const markForUser: IArticleAction = {
-                                label: gettext('Mark for user'),
-                                labelForGroup: gettext('Relations'),
-                                icon: 'icon-assign',
-                                onTrigger: () => {
-                                    manageMarkedUserForSingleArticle(superdesk, articleNext);
-                                },
-                            };
-
-                            const unmark: IArticleAction = {
-                                label: gettext('Unmark user'),
-                                labelForGroup: gettext('Relations'),
-                                icon: 'icon-assign',
-                                onTrigger: () => {
-                                    superdesk.entities.article.update({
-                                        ...articleNext,
-                                        marked_for_user: null,
-                                    });
-                                },
-                            };
-
-                            const markForOtherUser: IArticleAction = {
-                                label: gettext('Mark for other user'),
-                                labelForGroup: gettext('Relations'),
-                                icon: 'icon-assign',
-                                onTrigger: () => {
-                                    manageMarkedUserForSingleArticle(superdesk, articleNext);
-                                },
-                            };
-
-                            const assigned = articleNext.marked_for_user != null;
-
-                            if (assigned) {
-                                return Promise.resolve([unmark, markForOtherUser]);
-                            } else {
-                                return Promise.resolve([markForUser]);
-                            }
-                        },
-                        getActionsBulk: (_, articles: Array<IArticle>) => {
-                            if (articles.some(isPersonal)) {
-                                return Promise.resolve([]);
-                            }
-
-                            const selectedUserIds = uniq(
-                                articles
-                                    .map((item) => item.marked_for_user)
-                                    .filter((marked_for_user) => marked_for_user != null),
-                            );
-
-                            const initialUserId = selectedUserIds[0];
-                            const selectedUserIdInitial = selectedUserIds.length > 1 || initialUserId === null
-                                ? undefined
-                                : initialUserId;
-
-                            const message = selectedUserIds.length > 1
-                                ? gettext('Items are marked for different users')
-                                : undefined;
-
-                            return Promise.resolve([{
-                                label: gettext('Mark for user'),
-                                icon: 'icon-assign',
-                                onTrigger: () => {
-                                    superdesk.ui.showModal(getMarkForUserModal(
-                                        superdesk,
-                                        (selectedUserId) => {
-                                            articles.forEach((article) => {
-                                                superdesk.entities.article.update({
-                                                    ...article,
-                                                    marked_for_user: selectedUserId,
-                                                });
-                                            });
-                                        },
-                                        selectedUserIdInitial,
-                                        message,
-                                    ));
-                                },
-                            }]);
-                        },
+                        getActions: getActionsInitialize(superdesk),
+                        getActionsBulk: getActionsBulkInitialize(superdesk),
                     },
                 },
             },
